+++ src/proxy/utils.ts
@@ -1,11 +1,12 @@
-import { IncomingHttpHeaders } from 'http'
+import { IncomingHttpHeaders, IncomingMessage, ServerResponse } from 'http'
 import { Readable } from 'stream'
 import * as zlib from 'zlib'
 
-import { config } from '../config'
+import undici from 'undici'
+
 import { logger } from '../logger'
 
-import { IProxyRouteConfig } from './types'
+import { IProxyRouteConfig, IProxyRouteMethod } from './types'
 
 export const bufferToJson = (buffer: Buffer): unknown => {
 	try {
@@ -71,15 +72,91 @@ const stripPrefix = (reqUrl: string, prefix?: string): string => {
 }
 
 export const prepareHeaders = (
-	originalHeaders: IncomingHttpHeaders
+	originalHeaders: IncomingHttpHeaders,
+	remoteAddress?: string
 ): IncomingHttpHeaders => {
 	return {
 		...originalHeaders,
+		'X-Forwarded-For': originalHeaders['x-forwarded-for'] ?? remoteAddress,
 		// Next line bypasses: [ERR_TLS_CERT_ALTNAME_INVALID]: Hostname/IP does not match certificate's altnames.
-		...(config.get('env') === 'development' && { host: '' }),
+		host: '',
 	}
 }
 
 export const buildUrl = (route: IProxyRouteConfig, url: string): string => {
 	return `${route.targetHost}${stripPrefix(url, route.proxyPrefix)}`
+}
+
+export const pipeResponse = async (
+	url: string,
+	req: IncomingMessage,
+	res: ServerResponse
+): Promise<void> => {
+	await undici.stream(
+		url,
+		{
+			opaque: { url, res },
+			method: req.method as IProxyRouteMethod,
+			headers: prepareHeaders(req.headers, req.socket.remoteAddress),
+			// eslint-disable-next-line no-undefined
+			body: req.method === 'POST' ? req : undefined,
+		},
+		({ headers, opaque, statusCode }) => {
+			const { url: reqUrl, res: rawRes } = opaque as {
+				url: string
+				res: ServerResponse
+			}
+
+			logger.trace({ url: reqUrl, headers, statusCode }, 'Response')
+
+			rawRes.writeHead(statusCode, headers)
+
+			return rawRes
+		}
+	)
+}
+
+export const processBody = async (
+	route: IProxyRouteConfig,
+	url: string,
+	req: IncomingMessage,
+	res: ServerResponse
+): Promise<void> => {
+	// force compression to gzip and decompress the response
+	req.headers['accept-encoding'] = 'gzip'
+
+	const response = await undici.request(url, {
+		method: req.method as IProxyRouteMethod,
+		headers: prepareHeaders(req.headers, req.socket.remoteAddress),
+		// eslint-disable-next-line no-undefined
+		body: req.method === 'POST' ? req : undefined,
+	})
+
+	logger.trace(
+		{ url, headers: response.headers, statusCode: response.statusCode },
+		'Response'
+	)
+
+	if (
+		response.statusCode !== 200 ||
+		!response.headers['content-type']?.startsWith('application/json')
+	) {
+		res.writeHead(response.statusCode, response.headers)
+		response.body.pipe(res)
+		return
+	}
+
+	const body = await decompressResponse(response.body)
+	const json = bufferToJson(body)
+
+	if (json && route.jsonPostProcess) {
+		const finalJson = route.jsonPostProcess(json, req.headers.host)
+		const finalBody = await compressJson(finalJson)
+		response.headers['content-length'] = finalBody.length.toString()
+		res.writeHead(response.statusCode, response.headers)
+		res.end(finalBody)
+	} else {
+		res.writeHead(response.statusCode, response.headers)
+		res.end(body)
+	}
 }
